msm: camera: sensor: Add support for 3B read
Newer I2C slaves have 3bytes read address.
Add support to read the data from slaves that
support 3 byte address.

CRs-Fixed: 1013787

+++ drivers/media/platform/msm/camera_v2/sensor/cci/msm_cci.h
@@ -89,7 +89,7 @@ struct msm_camera_cci_gpio_cfg {
 };
 
 struct msm_camera_cci_i2c_read_cfg {
-	uint16_t addr;
+	uint32_t addr;
 	enum msm_camera_i2c_reg_addr_type addr_type;
 	uint8_t *data;
 	uint16_t num_byte;
